+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## [0.1.9] - Unreleased
+## [0.1.10] - Unreleased
+### Added
+* n/a
+### Fixed
+* n/a
+### Updated
+* n/a
+### Breaking changes
+* n/a
+
+## [0.1.9] - 2024-04-17
 ### Added
 * Support for table unions in table builder
 * New column property `raise_missing_arg_exception` to allow for some spark function inputs to be missing